Include route's element type module before function-specific ones.
This order makes more sense, not sure why this was like this before, but before the ElementMixin was included as a module ancestor which turns out to have lower priority (same as an actual ancestor class?) than a regular module. Ruby's method dispatch hierarchy gets complex enough at the edge that I suspect parts are simply undefined.