Fix my last changes to sort_section
The sorting window was offset by 1 element, causing odd results.

Clearing the to_sort array instead of creating a new empty one fixes it, I think because the to_sort object is resolved before starts.next is called, but it is a strange bug and may be new behaviour in JRuby 1.7.