Modifications to run with Bluprints, Pipes and Gremlin 2.4
Apparently most changes between 2.3 and 2.4 where classes being
renamed, moved from Gremlin to Pipes and some interfaces breaking.

PropertyFilterPipe constructor changed it's signature, using the
Predicate in the middle

FilterPipe::Filter was replaced by Compare

CollectionFilter now uses Contains as predicates instead of Compare

Most tests are passing, but I still have some issue that I believe
to be some bug on the CollectionFilter due to these changes, but
I'm not familiar enough with the codebase to troubleshoot it in
a reasonable time.